Convert from using a TUBE CHANNEL to a TEXT CHANNEL
Telepathy no longer supports TUBE CHANNEL (as of Fedora 20), so
collaboration breaks in activities that use TUBE. But TEXT CHANNEL is
still supported. This patch converts from TUBE to TEXT by swapping in
a helper class, TextChannelWrapper (textchannelwrapper.py is based on
the collaboration model used in the Chat activity).

Note that the collaboration model in this activity could have been
implemented in a TEXT CHANNEL to begin with as it uses no binary data
and it is a simple broadcast of data between all clients, so the
conversion was straight-forward.