Fix card image persistence
Images persisted in a new game.  Reproduce as follows; select the game with mouse, cheese and moon; play and finish game; start new game with just numbers or patterns.  Images from the earlier game, i.e mouse, cheese and moon are displayed again.

Cause is the deck of cards that are created at the beginning of the activity is used throughout. These cards have a parameter called Sprite which contains an _embedded_image which is set to mouse, cheese and moon images when that option is selected. However, for patterns, _embedded_image should ideally be None.

When the game is changed, the _embedded_image is not updated, and thus is rendered incorrectly.

Fix is to set self.spr each time a new game is created, so that the _embedded_image is set as and when necessary.